Refactoring the payable service and adding httpError package.

+++ payables/repository/interface.go
@@ -4,6 +4,6 @@ import "algogrit.com/yaes-server/entities"
 
 type PayableRepository interface {
 	RetrieveBy(entities.User) ([]*entities.Payable, error)
-	FindBy(uint64) (*entities.Payable, error)
+	FindBy(uint) (*entities.Payable, error)
 	Update(*entities.Payable) error
 }
